Add files via upload
I have provided the database along with all the other files. You have to save the database file into  C ->Users->Admin->Documents. And remember while you extract all the files check the properties so that and enable security unblock . You have to change the object name in the Program.cs file in the last to run a particular file. In this application, you can also print the bill for the user.  Any further queries you can mail me.